Fixing a few unit tests

+++ src/Microsoft.Owin.Compression/Storage/DefaultCompressedStorage.cs
@@ -210,7 +210,10 @@ namespace Microsoft.Owin.Compression.Storage
             {
                 if (Interlocked.Decrement(ref _references) == 0)
                 {
-                    File.Delete(PhysicalPath);
+                    if (File.Exists(PhysicalPath))
+                    {
+                        File.Delete(PhysicalPath);
+                    }
                 }
             }
         }
